refactor(core): shrink _core.py preamble into 3 focused modules (tier-10 PR-A)
Extract the _core.py preamble into three focused sibling modules:
- _core_error_injection.py: ERROR_INJECT_ENV_VAR, _get_error_injection_mode,
  _SyntheticErrorTransport (carries forward tier-9 A's env-var guard logic)
- _core_constants.py: DEFAULT_TIMEOUT, DEFAULT_CONNECT_TIMEOUT,
  DEFAULT_KEEPALIVE_MIN_INTERVAL, DEFAULT_MAX_CONCURRENT_UPLOADS,
  DEFAULT_MAX_CONCURRENT_RPCS
- _core_helpers.py: AUTH_ERROR_PATTERNS, is_auth_error,
  _resolve_keepalive_interval

_decode_response_late_bound and _sleep_late_bound deliberately STAY in
_core.py because RpcExecutor lambdas resolve them via _core's module
globals at call time, and monkeypatch.setattr("notebooklm._core.decode_response", ...)
tests depend on that resolution path.

MAX_RETRY_AFTER_SECONDS stays canonical in _core_transport.py:22 and is
re-imported by _core.py as before.

Stale-comment sweep: _core_lifecycle.py:14, :145, :233-234 and
_core.py:620 updated to reference new module locations.

Permanent assertion added to tests/unit/test_public_shims.py confirming
notebooklm._core.is_auth_error resolves through _core_helpers.

_core.py: 1587 -> 1323 LOC (-264).
