+++ src/notebooklm/_core_helpers.py
@@ -1,0 +1,91 @@
+"""Cross-seam helpers: auth-error classification, keepalive validation.
+
+Small, pure helpers extracted from :mod:`notebooklm._core`'s preamble. Each
+helper is re-exported from :mod:`notebooklm._core` so existing
+``from notebooklm._core import is_auth_error`` imports keep working unchanged.
+
+These helpers stay separate from :mod:`notebooklm._core_constants` because
+they carry behavior (and therefore tests), while the constants module is
+data-only.
+"""
+
+from __future__ import annotations
+
+import math
+
+import httpx
+
+from .rpc import (
+    AuthError,
+    ClientError,
+    NetworkError,
+    RateLimitError,
+    RPCError,
+    RPCTimeoutError,
+    ServerError,
+)
+
+# Auth error detection patterns (case-insensitive)
+AUTH_ERROR_PATTERNS = (
+    "authentication",
+    "expired",
+    "unauthorized",
+    "login",
+    "re-authenticate",
+)
+
+
+def _resolve_keepalive_interval(keepalive: float | None, min_interval: float) -> float | None:
+    """Validate and clamp the keepalive interval.
+
+    ``None`` disables the background task. Otherwise both values must be
+    positive finite numbers; the effective interval is ``max(keepalive,
+    min_interval)`` so callers can't accidentally lower the rate-limit floor.
+    """
+    if not (math.isfinite(min_interval) and min_interval > 0):
+        raise ValueError(
+            f"keepalive_min_interval must be a positive finite number, got {min_interval!r}"
+        )
+    if keepalive is None:
+        return None
+    if not (math.isfinite(keepalive) and keepalive > 0):
+        raise ValueError(f"keepalive must be None or a positive finite number, got {keepalive!r}")
+    return max(keepalive, min_interval)
+
+
+def is_auth_error(error: Exception) -> bool:
+    """Check if an exception indicates an authentication failure.
+
+    Args:
+        error: The exception to check.
+
+    Returns:
+        True if the error is likely due to authentication issues.
+    """
+    # AuthError is always an auth error
+    if isinstance(error, AuthError):
+        return True
+
+    # Don't treat network/rate limit/server errors as auth errors
+    # even if they're subclasses of RPCError
+    if isinstance(
+        error,
+        NetworkError | RPCTimeoutError | RateLimitError | ServerError | ClientError,
+    ):
+        return False
+
+    # HTTP 400/401/403 are auth errors.
+    # Google returns 400 for expired CSRF tokens (not 401/403). Layer-1
+    # recovery (refresh_auth) re-extracts SNlM0e from the NotebookLM
+    # homepage and retries with a fresh token. The retry guard
+    # (``_is_retry`` in ``rpc_call``) bounds wasted refreshes on legitimate
+    # 400s (bad payload) to one extra GET per call.
+    if isinstance(error, httpx.HTTPStatusError):
+        return error.response.status_code in (400, 401, 403)
+
+    # RPCError with auth-related message
+    if isinstance(error, RPCError):
+        message = str(error).lower()
+        return any(pattern in message for pattern in AUTH_ERROR_PATTERNS)
+
+    return False
